Split cutscenes into their own files
